+++ src/index.js
@@ -10,7 +10,7 @@ import Proxy from './utils/Proxy'
 import Logger from './utils/Logger'
 import FileSystem from './utils/FileSystem'
 import Contracts from './utils/Contracts'
-import Deploy from './utils/Deploy'
+import { sendTransaction, deploy } from './utils/Transactions'
 
 // test behaviors
 import { behaviors, helpers } from './test'
@@ -29,11 +29,12 @@ export {
   assertRevert,
   assertions,
   behaviors,
+  sendTransaction,
+  deploy,
   Proxy,
   Logger,
   FileSystem,
   Contracts,
-  Deploy,
   App,
   Package,
   Release,